feat(frontend): consolidate ServiceBay updater + service image-updates on Home
Bundle both update surfaces into one coherent "Updates" area on the Home
overview: the ServiceBay self-updater (version status + Check Now + Update Now,
GET/POST /api/system/update) sits next to the per-service image-updates banner
(#2069/4.140.0), each with its own trigger button, under a single SectionHeading.

Extracted the self-updater half of UpdatesSection into a shared
ServiceBayUpdateCard component (also owns the AppUpdateStatus type, re-exported
from settings helpers for back-compat) so Settings and Home share one updater
without duplicating the endpoint logic. The OS-reinstall block stays in
UpdatesSection (settings-only, destructive/recovery — not Home material).

First real consumer of the new components/ui SectionHeading primitive.

Closes #2082
